requirements.txt
selenium==4.19.0
pytest==8.1.1
allure-pytest==2.13.5
pytest-xdist==3.6.1
requests==2.32.3
jsonschema==4.23.0
pydantic==2.8.2